requirements.txt
Django>=5.2,<5.3
django-ninja>=1.0
psycopg2-binary>=2.9
celery>=5.3
redis>=5.0
python-dotenv>=1.0
